Removed dependency from 52n-wps-ags to 52-wps-mc. Migrated some old classes from the 52-wps-mc module to the 52n-wps-ags
