--- 52n-wps-mc/src/main/java/org/n52/wps/server/feed/movingcode/AlgorithmDescription.xsd
@@ -1,110 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" elementFormDefault="qualified" attributeFormDefault="qualified">
-	<xs:element name="AlgorithmDescription">
-		<xs:annotation>
-			<xs:documentation>Deployment Description for a WPS Algorithm</xs:documentation>
-		</xs:annotation>
-		<xs:complexType>
-			<xs:sequence>
-				<xs:element name="workspaceLocation" type="xs:anyURI">
-					<xs:annotation>
-						<xs:documentation>URL pointing to the workspace root, e.g. "file://level1/level2/workspaceX"</xs:documentation>
-					</xs:annotation>
-				</xs:element>
-				<xs:element name="algorithmLocation" type="xs:anyURI">
-					<xs:annotation>
-						<xs:documentation>URL pointing to the algorithm; relative to the workspace root. If appropriate, this location can point to some location within the container, e.g. "algorithm://executables/tools.tbx?mytools/tool2"</xs:documentation>
-					</xs:annotation>
-				</xs:element>
-				<xs:element name="containerType" type="xs:anyURI">
-					<xs:annotation>
-						<xs:documentation>URN referencing the specific container, e.g. urn:n52:wps:algorithmcontainer:arctoolbox:9.3</xs:documentation>
-					</xs:annotation>
-				</xs:element>
-				<xs:element name="requiredRuntimeComponent" type="xs:anyURI" maxOccurs="unbounded">
-					<xs:annotation>
-						<xs:documentation>URN referencing a required processing backend, e.g. urn:n52:wps:gpsystem:arcgis:9.3</xs:documentation>
-					</xs:annotation>
-				</xs:element>
-				<xs:element name="algorithmParameters">
-					<xs:annotation>
-						<xs:documentation>A collection of the algorithm parameters</xs:documentation>
-					</xs:annotation>
-					<xs:complexType>
-						<xs:sequence>
-							<xs:element name="parameter" type="AlgorithmParameterType" maxOccurs="unbounded">
-								<xs:annotation>
-									<xs:documentation>The tool's parameters</xs:documentation>
-								</xs:annotation>
-							</xs:element>
-							<xs:element name="separatorString" type="xs:string" minOccurs="0">
-								<xs:annotation>
-									<xs:documentation>A string that acts as a seperator between parameters, e.g. " " for SPACE</xs:documentation>
-								</xs:annotation>
-							</xs:element>
-						</xs:sequence>
-						<xs:attribute name="sequential" type="xs:boolean">
-							<xs:annotation>
-								<xs:documentation>If true, this tool's parameters are indexed through their position as integers. Otherwise, the mapping from the ows:Identifier to the respective LegacyID is done via name strings.</xs:documentation>
-							</xs:annotation>
-						</xs:attribute>
-					</xs:complexType>
-				</xs:element>
-			</xs:sequence>
-		</xs:complexType>
-	</xs:element>
-	<!-- ========================================================================== -->
-	<xs:complexType name="AlgorithmParameterType">
-		<xs:annotation>
-			<xs:documentation>Structure for a parameter description.</xs:documentation>
-		</xs:annotation>
-		<xs:sequence>
-			<xs:element name="prefixString" type="xs:string" minOccurs="0">
-				<xs:annotation>
-					<xs:documentation>A prefix string that shall be added before this parameter, e.g. "-p("</xs:documentation>
-				</xs:annotation>
-			</xs:element>
-			<xs:element name="suffixString" type="xs:string" minOccurs="0">
-				<xs:annotation>
-					<xs:documentation>A suffix string that shall be added before this parameter, e.g. ")"</xs:documentation>
-				</xs:annotation>
-			</xs:element>
-			<xs:element name="separatorString" type="xs:string" minOccurs="0">
-				<xs:annotation>
-					<xs:documentation>A string that acts as a separator between individual parameter values, e.g. " " for SPACE</xs:documentation>
-				</xs:annotation>
-			</xs:element>
-			<xs:choice>
-				<xs:annotation>
-					<xs:documentation>The Legacy ID by which the parameters are defined in the Legacy environment. Use legacyIntID for sequential parameters, legacyStrindID for parameters name strings</xs:documentation>
-				</xs:annotation>
-				<xs:element name="positionID" type="PositionIDType"/>
-				<xs:element name="stringID" type="StringIDType"/>
-			</xs:choice>
-			<xs:element name="wpsInputID" type="xs:string" minOccurs="0">
-				<xs:annotation>
-					<xs:documentation>The WPS Input ID this parameter shall be mapped to.</xs:documentation>
-				</xs:annotation>
-			</xs:element>
-			<xs:element name="wpsOutputID" type="xs:string" minOccurs="0">
-				<xs:annotation>
-					<xs:documentation>The WPS Output ID this parameter shall be mapped to.</xs:documentation>
-				</xs:annotation>
-			</xs:element>
-		</xs:sequence>
-	</xs:complexType>
-	<!-- ========================================================================== -->
-	<xs:simpleType name="PositionIDType">
-		<xs:annotation>
-			<xs:documentation>Positive integer indicationg the position of this parameter.</xs:documentation>
-		</xs:annotation>
-		<xs:restriction base="xs:positiveInteger"/>
-	</xs:simpleType>
-	<!-- ========================================================================== -->
-	<xs:simpleType name="StringIDType">
-		<xs:annotation>
-			<xs:documentation>String indicating the name of this parameter.</xs:documentation>
-		</xs:annotation>
-		<xs:restriction base="xs:string"/>
-	</xs:simpleType>
-</xs:schema>
