--- 52n-wps-mc/src/main/java/org/n52/wps/server/feed/movingcode/ObjectFactory.java
@@ -1,91 +1,0 @@
-/**
- * ﻿Copyright (C) 2012 - 2014 52°North Initiative for Geospatial Open Source
- * Software GmbH
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 as published
- * by the Free Software Foundation.
- *
- * If the program is linked with libraries which are licensed under one of
- * the following licenses, the combination of the program with the linked
- * library is not considered a "derivative work" of the program:
- *
- *       • Apache License, version 2.0
- *       • Apache Software License, version 1.0
- *       • GNU Lesser General Public License, version 3
- *       • Mozilla Public License, versions 1.0, 1.1 and 2.0
- *       • Common Development and Distribution License (CDDL), version 1.0
- *
- * Therefore the distribution of the program linked with libraries licensed
- * under the aforementioned licenses, is permitted by the copyright holders
- * if the distribution is compliant with both the GNU General Public
- * License version 2 and the aforementioned licenses.
- *
- * This program is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details.
- */
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.3-hudson-jaxb-ri-2.2.3-3-
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a>
-// Any modifications to this file will be lost upon recompilation of the source schema.
-// Generated on: 2011.02.08 at 05:49:57 PM MEZ
-//
-
-
-package org.n52.wps.server.feed.movingcode;
-
-import javax.xml.bind.annotation.XmlRegistry;
-
-
-/**
- * This object contains factory methods for each
- * Java content interface and Java element interface
- * generated in the org.n52.wps.server.feed.schema package.
- * <p>An ObjectFactory allows you to programatically
- * construct new instances of the Java representation
- * for XML content. The Java representation of XML
- * content can consist of schema derived interfaces
- * and classes representing the binding of schema
- * type definitions, element declarations and model
- * groups.  Factory methods for each of these are
- * provided in this class.
- *
- */
-@XmlRegistry
-public class ObjectFactory {
-
-
-    /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: org.n52.wps.server.feed.schema
-     *
-     */
-    public ObjectFactory() {
-    }
-
-    /**
-     * Create an instance of {@link AlgorithmDescription }
-     *
-     */
-    public AlgorithmDescription createAlgorithmDescription() {
-        return new AlgorithmDescription();
-    }
-
-    /**
-     * Create an instance of {@link AlgorithmDescription.AlgorithmParameters }
-     *
-     */
-    public AlgorithmDescription.AlgorithmParameters createAlgorithmDescriptionAlgorithmParameters() {
-        return new AlgorithmDescription.AlgorithmParameters();
-    }
-
-    /**
-     * Create an instance of {@link AlgorithmParameterType }
-     *
-     */
-    public AlgorithmParameterType createAlgorithmParameterType() {
-        return new AlgorithmParameterType();
-    }
-
-}
